#98: move respective test cases from ProductVariantUpdateActionUtilsTest to ProductUpdateActionUtilsTest
